Bug 1211568 - Update NSS to 3.21 Beta 3, r=kaie
UltraBlame original commit: 5dc991cc398a67f0bdaddcd923eceae1b5a2bc5d

+++ security/nss/lib/ssl/ssl3gthr.c
@@ -71,8 +71,8 @@ ssl3_GatherData(sslSocket *ss, sslGather *gs, int flags)
 	    break;
 	}
 
-	PORT_Assert( nb <= gs->remainder );
-	if (nb > gs->remainder) {
+	PORT_Assert( (unsigned int)nb <= gs->remainder );
+	if ((unsigned int)nb > gs->remainder) {
 	    
 	    gs->state = GS_INIT;         
 	    rv = SECFailure;
